moved things around and split up WorkspaceSummaryPage
