+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/component/Component.scala
@@ -1,6 +1,6 @@
-package org.broadinstitute.dsde.firecloud.page.components
+package org.broadinstitute.dsde.firecloud.component
 
-import org.broadinstitute.dsde.firecloud.page.FireCloudView
+import org.broadinstitute.dsde.firecloud.FireCloudView
 import org.openqa.selenium.WebDriver
 
 abstract class Component(id: String)(implicit webDriver: WebDriver) extends FireCloudView {
