+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/component/TabBar.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.dsde.firecloud.page.components
+package org.broadinstitute.dsde.firecloud.component
 
 import org.openqa.selenium.WebDriver
