+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/user/SignInPage.scala
@@ -1,6 +1,7 @@
 package org.broadinstitute.dsde.firecloud.page.user
 
-import org.broadinstitute.dsde.firecloud.page.{FireCloudView, PageUtil}
+import org.broadinstitute.dsde.firecloud.FireCloudView
+import org.broadinstitute.dsde.firecloud.page.PageUtil
 import org.broadinstitute.dsde.firecloud.test.WebBrowserUtil
 import org.openqa.selenium.WebDriver
 import org.scalatest.selenium.{Page, WebBrowser}
