+++ automation/src/test/scala/org/broadinstitute/dsde/firecloud/page/workspaces/summary/DeleteWorkspaceModal.scala
@@ -1,0 +1,33 @@
+package org.broadinstitute.dsde.firecloud.page.workspaces.summary
+
+import org.broadinstitute.dsde.firecloud.FireCloudView
+import org.openqa.selenium.WebDriver
+
+/**
+  * Page class for the workspace delete confirmation modal.
+  */
+class DeleteWorkspaceModal(implicit webDriver: WebDriver) extends FireCloudView {
+
+  /**
+    * Confirms the request to delete a workspace. Returns after the FireCloud
+    * busy spinner disappears.
+    */
+  def confirmDelete(): Unit = {
+    ui.clickConfirmDeleteButton()
+  }
+
+  def confirmDeleteWait(): Unit = {
+    // Micro-sleep to make sure the spinner has had a chance to render
+    Thread sleep 200
+    await notVisible spinner
+  }
+
+
+  object ui {
+    private val confirmDeleteButtonQuery: Query = testId("confirm-delete-workspace-button")
+
+    def clickConfirmDeleteButton(): Unit = {
+      click on (await enabled confirmDeleteButtonQuery)
+    }
+  }
+}
